Add phpunit workflow
Fix tests

ERM44815
